policy: repository updates caches within rules
* Update `ResolvePolicy` to take in a `SecurityIdentity` instead of just the
`LabelArray` within a `SecurityIdentity` so that the `EndpointsSelected` field
within each `rule` can be kept up-to-date in relation to this endpoint. Update
several related, unexported functions within the policy package to account for
this as well.
* Update `AddListLocked`, `DeleteByLabelsLocked`, to return the slice of `rule`
which was added. This slice will be utilized to update the caches within each
`rule` which specify what endpoints the rule selects. The analysis is performed
in relation to these rules so that it can be determined which set of endpoints
on the node need to be regenerated, as the rules which were added or deleted may
not select all endpoints.

Signed-off by: Ian Vernon <ian@cilium.io>