serialize policy repository update, endpoint events
Previously, when policy updates occurred, all endpoints were regenerated,
regardless of whether or not the rules which were updated, added, or deleted
actually selected all endpoints. This commit removes this kitchen-sink approach
to policy calculation when the repository is modified, and instead regenerates
only the endpoints that are selected by the rules that are updated, added, or
deleted. If an endpoint is not selected by the changes in the policy repository,
then its policy revision number is bumped. These events must be synchronized /
mutually exclusive in their operation, so the endpoint's EventQueue is utilized
to enqueue revision bump events and regeneration events. This is because an
Endpoint's policy revision should not be bumped while it is being regenerated,
as this could signal that the Endpoint is realizing a policy revision which it
hasn't yet (as it is still regenerating!).

Now that endpoints are selectively regenerated per change in the policy
repository, the ordering of when all work associated with a given policy update
occurs is important. Take the following example:

* Time X: Policy imported with rules that select endpoint A. Eventually, this
endpoint will be regenerated by having a regeneration event pushed onto its
event queue, and this may be done after the policy has been added, as it is done
in an asynchronous manner.
* Time X+1: Policy imported with rules that do not select endpoint A.
Eventually, this endpoint will not be regenerated, and instead will have its
policy revision incremented, after the API has returned, similar to Time X.
However, because the regeneration triggered at Time X and the policy revision
incrementing at Time X+1 are launched as goroutines, there is no guarantee
about when either event will be enqueued for endpoint A. This may mean that the
policy revision event will get enqueued *before* the regeneration event for
the endpoint, which means that the endpoint will falsely report that it is
implementing that policy revision, while it has not yet because the regeneration
for the rules from Time X has not occurred yet.

This means that there needs to be a way to ensure that updating of the policy
repository, and the caches within the policy repository rules indicating which
endpoints the rules select, is done in-order. Two EventQueues are added: one,
which serializes the processing of policy repository changes (add, delete,
update), and one which serializes the 'reacting' to rule updates (e.g., which
manages the enqueuing of events for endpoints in an in-order fashion).

Signed-off by: Ian Vernon <ian@cilium.io>